# swagger/swagger.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: HomeNet
  description: |
    listings of properties for sale
  license:
    name: MIT
servers:
- url: http://localhost:3000
paths:
...
  /listings/_id:
      get:
        summary: gets a post by id
        operationId: getById
        tags:
        - listings
        parameters:
        - name: _id
          in: path
          description: the post by id
          required: true
          schema:
            type: string
        responses:
          '200':
            description: the post
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/ResponseListings"
          default:
            description: unexpected error
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/Error"
      put:
        summary: updates a post
        operationId: put
        tags:
        - listings
        parameters:
        - name: _id
          in: path
          description: the post
          required: true
          schema:
            type: string
        requestBody:
          description: the post to be updated
          required: true
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseListings'
        responses:
          '200':
            description: the replaced post
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/ResponseListings"
          default:
            description: unexpected error
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/Error"
      delete:
        summary: deletes a post
        operationId: deleteing
        tags:
          - listings
        parameters:
        - name: _id
          in: path
          description: id
          required: true
          schema:
            type: string
        responses:
          '200':
            description: the deleted post
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/ResponseListings"
          '204':
            description: No change (nothing was deleted)
          default:
            description: unexpected error
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/Error"
components:
  schemas:
    Listing:
      type: object
      properties:
        latitude:
          type: number
        longitude :
          type: number
        street:
          type: string
        city:
          type: string
        municipality:
          type: string
        summary:
          type: string
        price:
          type: number
        monthlyFee:
          type: number
        bidding:
          type: boolean
    ResponseListing:
      allOf:
      - $ref: "#/components/schemas/Listing"
      - properties:
          __v:
            type: integer
            format: uint32
          _id:
            type: string
            format: uuid
    ResponseListings:
      type: array
      items:
        $ref: "#/components/schemas/ResponseListing"
    Error:
      required:
        - error
      properties:
        error:
          type: object
          properties:
            expose:
              type: boolean
            statusCode:
              type: integer
              format: uint32
            status:
              type: integer
              format: uint32
            body:
              title: body as escaped json
              type: string
            type:
              title: fault desciption
              type: string